+++ src/main/java/baseball/Application.java
@@ -1,10 +1,10 @@
 package baseball;
 
-import baseball.computer.Computer;
+import baseball.game.Computer;
 import baseball.game.Game;
 import baseball.game.GameRestartPolicy;
 import baseball.game.GameStarter;
-import baseball.player.Player;
+import baseball.game.Player;
 
 public class Application {
     public static void main(String[] args) {
