+++ src/main/java/baseball/context/ResultMessage.java
@@ -1,4 +1,4 @@
-package baseball.computer;
+package baseball.context;
 
 public class ResultMessage {
     private String ball;
