+++ src/main/java/baseball/game/Game.java
@@ -1,10 +1,8 @@
 package baseball.game;
 
-import baseball.computer.Computer;
 import baseball.context.GameCycle;
 import baseball.context.Numbers;
 import baseball.context.RestartPolicy;
-import baseball.player.Player;
 
 public class Game implements GameCycle {
