+++ src/main/java/baseball/game/GameRestartPolicy.java
@@ -1,8 +1,8 @@
 package baseball.game;
 
-import baseball.context.Printer;
+import baseball.io.Printer;
 import baseball.context.RestartPolicy;
-import baseball.printer.ConsolePrinter;
+import baseball.io.ConsolePrinter;
 import camp.nextstep.edu.missionutils.Console;
 
 public class GameRestartPolicy implements RestartPolicy {
